Reapply "[Member] 회원가입 휴대폰 인증 완료"
This reverts commit 1e8f7f25b54c62da087351496c1639b11970094b.

+++ src/main/java/com/eob/shop/model/data/ShopEntity.java
@@ -175,10 +175,10 @@ public class ShopEntity {
     /**
      * 상점 위치 - 경도
      */
-    private float longitude;
+    private Double longitude;
 
     /**
      * 상점 위치 - 위도
      */
-    private float latitude;
+    private Double latitude;
 }
